Mostly fix general format.
2002-03-26  Morten Welinder  <terra@diku.dk>

	* src/rendered-value.c (rendered_value_new): Just peek directly
	into the approx_width structure -- everyone else does that.  Take
	worst case of "e", signs, and decimal points into account.

	* src/style.h (StyleFont): Change floats to doubles.  Add more
	approx_width fields.

	* src/style.c (style_font_get_width_pts): Remove.

+++ src/format.h
@@ -36,7 +36,7 @@ StyleFormat   *style_format_default_money	(void);
 
 void   format_destroy (StyleFormat *format);
 char  *format_value   (StyleFormat const *format, Value const *value, StyleColor **color,
-		       float col_width);
+		       double col_width);
 
 void   format_color_init     (void);
 void   format_color_shutdown (void);
